Revert "Use a flowtable."
I was getting way worse performance with the flowtable, and "Error:
Could not process rule: No such file or directory" when nftables tried
to add the flowtable at boot. Maybe I could debug it a bit to figure out
what's going on, but that's probably only worth it if I notice
performance problems without using flowtable.

This reverts commit 4e69793cde343731db431ae937047d2f66e39938.