Configure partitions for kubernetes+ceph nodes
I couldn't find any way to really control the size of an encrypted LVM
partition using partman, so with this I can do it in three stages.
Configure outer partitions with partman, run
make-encrypted-system-partition from a shell, configure the logical root
volume with partman. I also don't see an obvious way to get partman to
write changes to disk after the first stage, but selecting the option to
setup LVM is a non-obvious way to do that.

If I later do find a way to control the size with partman, I think the
passdev parts of this can be done with a partman.early_command.d script
to set up the key file and preseed the passphrase from that file, and a
preseed.late_command.d script to modify the crypttab line generated by
partman.