Use salt.states.cron instead of cron files
The major benefit seems to be the ability to use 'random' to spread jobs
out at different times on different minions.

salt.states.cron uses user crontabs instead of system-wide ones, which
means the disk_usage cron job that ran as nobody would have been able to
be modified by nobody. To avoid giving nobody permissions on anything,
and because that cron job should be safe to run as root, I switched it
to run as root.