Always explicitly set IPv6AcceptRA
On a router, I thought the default value would be reasonable:
> If unset, the kernel's default is used, and RAs are accepted only when
> local forwarding is disabled for that interface.

However, from looking at the systemd-networkd source code (the latest
version, not the version running on any system I'm using this config), I
*think* it actually checks the value of the IPForward config item,
rather than the kernel's idea about whether forwarding is enabled or
not. Explicitly setting IPv6AcceptRA avoids the issue entirely, and also
does the right thing (according to the data in the pillar) when
forwarding isn't enabled.

Commit 1368d99dff970642c2909043e97ff78be9bc52ae talked about a loop
between systemd-networkd responding to RAs and dnsmasq responding to
changes in the network interface's configuration. That commit fixed a
tight loop that was pegging the CPU, but I've also been noticing
intermittent IPv6 connectivity issues between non-router hosts on the
same segment.

Looking into those issues more, it looked like each host had a /128
assigned by DHCPv6 and a /64 from SLAAC. On the host side, the
connectivity issues seemed to happen when the /64 was missing, so the
default route to the router caught any traffic to other addresses on the
same segment. On the router side, it looked like in addition to the
router's intended static /64, it also had a SLAAC /64 most-but-not-all
of the time. And the connectivity issues (which did not impact the
router itself as far as I could tell) corresponded with the times the
router did not have the SLAAC /64. My theory is that systemd-networkd
was accepting RAs from dnsmasq to configure the SLAAC /64. Then when
that timed out, dnsmasq was somehow removing the entire /64 from its
list of prefixes to advertise, so other hosts lost the /64 too. But then
after some time things started working again, I'm guessing because of a
periodic re-check of something in either systemd-networkd or dnsmasq.

Commit 1368d99dff970642c2909043e97ff78be9bc52ae also said that
IPv6AcceptRA=false didn't work. I'm not sure, but I suspect that I
didn't adequately restart systemd-networkd after trying out
IPv6AcceptRA=false, because it seems to work now after rebooting the
computer. The router no longer shows SLAAC addresses on the static
interfaces, `ip monitor address` on the router no longer shows any
periodic changes to the static interfaces, and pinging over IPv6 between
hosts on the same segment showed 0.0066012% packet loss over almost 17
hours. (One side was a laptop on wifi, so I suspect the wifi was to
blame for the 4 lost packets.)