Merge local_relay and postfix states back into mail
With postmulti, it looks like there needs to be a single main instance,
and I think the local relay one should be it, because it's the only one
that all hosts will have.

Also, use postalias with a macro to make it easier to run postalias (and
also postmap) with other files too.