Add names for wireguard interface addresses
I just had a pretty confusing broken ssh connection that this would have
helped with. As far as I can tell, I had ssh'ed through a wireguard
interface to the address associated with a bridge interface. Then when I
removed the last member of that bridge, the bridge interface lost its
address and the ssh connection froze. That also happened to be the only
non-wireguard home_router interface on that host, so it returned no IP
addresses for its name.

This change might not prevent a frozen ssh connection in that scenario,
but it at least makes it easier to see that the host is still responding
to DNS queries and still available on the wireguard interface(s). And it
makes it easier to immediately reconnect, automatically using one of the
remaining names from a wireguard interface.