Merge auth-related config into one .conf file
The split between passwd and oauth2 wasn't great. There were things in
the passwd conf file that were unrelated to the passwd files. Also, the
oauth2 conf file added auth_mechanisms that affected the passwd config.
By merging the two, it's more clear how the auth stuff all works, and
the passwd passdb won't be used for the oauth2 mechanisms.