Replace grains.filter_by with local jinja code
salt-ssh seems to run execution module functions (like grains.filter_by)
in serial while templating sls files, and that seems to account for the
vast majority of the time that salt-ssh is taking for me. This change
reduced the time to run state.test on one host from ~45s to ~30s for me.