Add a macro to generate cron lines in one file
With the cron.present and cron.absent states, it's possible to remove
old/unexpected cron jobs, but it requires a bunch of manual checking for
them. Those checks use execution modules, which are really slow in
salt-ssh. This avoids that by accumulating lines in one file, so when a
cron job state is removed, the next update of that file won't have the
old line.

Also, this makes it relatively easy to implement
https://github.com/saltstack/salt/issues/60567 while I'm at it.