Update dovecot config for 2.4
10-auth.conf moved from the SLS file to its own template because now it
includes the oauth2 client secret directly, and that uses file.read on a
file that isn't necessarily available when the SLS file is rendered.

I'm not sure what exactly changed, but oauth2_force_introspection seems
to be needed now. Without it, dovecot thinks the list of scopes is
empty, so it gives an error about the missing email-read-write scope.

Since 10-auth.conf is no longer world-readable (see above),
mail-storage-monitor-subscriptions now runs as root.