Set music_directory explicitly
mpd recently stopped working, because it was looking for foo/bar in
~/foo/bar instead of ~/Music/foo/bar. I'm not sure exactly what
happened, but ~/Music is a symlink to a subfolder of a network mount,
and the problem was soon after the symlink was broken because of a
broken network connection. Maybe mpd defaults to ~ instead of ~/Music
when ~/Music is a broken symlink, but I'm not sure. Either way, this
seems to work.